requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
opencv-python==4.8.1.78
redis==5.0.1
aiohttp==3.9.1
asyncio==3.4.3
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6